refactor(gateway): extract GatewayMediaMixin from run.py (slice 22 of #54962)
